+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Collections/Dictionaries became non null when at all possible
 - Class creation methods become Swift inits
 - Deprecated `FBSDKLoginBehaviorSystemAccount`
+- Deprecated `FBSDKLoginBehaviorWeb`
 - Deprecated `[FBSDKLoginManager renewSystemCredentials] `
 
 ### Removed